Decode names correctly
According
to: (https://github.com/noctarius/lightify-binary-protocol#basics-about-the-protocol)
strings are transmitted with utf-8 encoded. I could verify, that this
helps with a name of one of my groups, that contains a `ü`, and is now
correctly decoded. I did not test this for an ordinary light, but I
expect it to work as well.